requirements.txt
alembic==1.11.1
aniso8601==9.0.1
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.2
click==8.1.3
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Migrate==2.1.1
flask-restplus==0.13.0
flask-restx==1.1.0
Flask-Script==2.0.5
Flask-SQLAlchemy==3.0.3
Flask-Testing==0.8.1
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
Mako==1.2.4
MarkupSafe==2.1.2
PyJWT~=2.3.0
pyrsistent==0.19.3
pytz==2023.3
six==1.16.0
SQLAlchemy~=2.0.13
typing_extensions==4.5.0
Werkzeug~=2.3.3
psycopg2-binary==2.9.6
python-dotenv==1.0.0
cryptography==40.0.2
marshmallow~=3.19.0
Scrapy~=2.9.0
selenium~=4.9.1
itemadapter~=0.8.0
crochet~=2.0.0
scrapy-selenium==0.0.7
webdriver-manager==3.8.6
